+++ js/config/configs/regex/no-html-entities-jsx.js
@@ -8,7 +8,7 @@ module.exports = {
       'error', [
         {
           message: 'Do not use HTML Entities, Use UTF-8 instead',
-          regex: '&(#\\d{1,5}|[a-zA-Z0-9]{1,8});',
+          regex: '&(#\\d{1,5})|(#x[\\da-fA-F]{1,5})|([a-zA-Z0-9]{1,8});',
           files: {
             inspect: '\\.[jt]sx'
           }